requirements.txt
pdfplumber~=0.11.9
pytesseract~=0.3.13
pdf2image~=1.17.0
python-dotenv
langchain
langchain-community
langchain-openai
faiss-cpu
rank-bm25
docx2pdf